+++ polonius-engine/src/output/mod.rs
@@ -14,6 +14,7 @@ use std::collections::{BTreeMap, BTreeSet};
 
 mod datafrog_opt;
 mod hybrid;
+mod initialization;
 mod liveness;
 mod location_insensitive;
 mod naive;
@@ -61,7 +62,15 @@ impl ::std::str::FromStr for Algorithm {
 }
 
 #[derive(Clone, Debug)]
-pub struct Output<Region: Atom, Loan: Atom, Point: Atom, Variable: Atom> {
+pub struct Output<Region, Loan, Point, Variable, MovePath>
+where
+    Region: Atom,
+    Region: Atom,
+    Loan: Atom,
+    Point: Atom,
+    Variable: Atom,
+    MovePath: Atom,
+{
     pub errors: FxHashMap<Point, Vec<Loan>>,
 
     pub dump_enabled: bool,
@@ -76,6 +85,8 @@ pub struct Output<Region: Atom, Loan: Atom, Point: Atom, Variable: Atom> {
     pub subset_anywhere: FxHashMap<Region, BTreeSet<Region>>,
     pub var_live_at: FxHashMap<Point, Vec<Variable>>,
     pub var_drop_live_at: FxHashMap<Point, Vec<Variable>>,
+    pub path_maybe_initialized_at: FxHashMap<Point, Vec<MovePath>>,
+    pub var_maybe_initialized_on_exit: FxHashMap<Point, Vec<Variable>>,
 }
 
 /// Compares errors reported by Naive implementation with the errors
@@ -114,15 +125,16 @@ fn compare_errors<Loan: Atom, Point: Atom>(
     differ
 }
 
-impl<Region, Loan, Point, Variable> Output<Region, Loan, Point, Variable>
+impl<Region, Loan, Point, Variable, MovePath> Output<Region, Loan, Point, Variable, MovePath>
 where
     Region: Atom,
     Loan: Atom,
     Point: Atom,
     Variable: Atom,
+    MovePath: Atom,
 {
     pub fn compute(
-        all_facts: &AllFacts<Region, Loan, Point, Variable>,
+        all_facts: &AllFacts<Region, Loan, Point, Variable, MovePath>,
         algorithm: Algorithm,
         dump_enabled: bool,
     ) -> Self {
@@ -162,6 +174,8 @@ where
             subset_anywhere: FxHashMap::default(),
             var_live_at: FxHashMap::default(),
             var_drop_live_at: FxHashMap::default(),
+            path_maybe_initialized_at: FxHashMap::default(),
+            var_maybe_initialized_on_exit: FxHashMap::default(),
             dump_enabled,
         }
     }